requirements.txt
scrapy>=2.8.0
scrapy-playwright>=0.1.0
python-dateutil>=2.8.0